refactor(composable): Move form validator to dedicated file
- Move `useFormValidator` from `utils/composable.ts` to `composables/form-validator.ts`
- Remove unused imports in `ReqNewCertDlg.vue` and `SignIn.vue`
- Clean up `utils/composable.ts` by removing form validator logic
